requirements.txt
# Core ML/Data Science packages
torch==2.6.0
numpy==1.25.2
pandas==2.2.3
matplotlib==3.9.4
scikit-learn==1.6.1
tqdm==4.67.1

# NLP and Transformers
transformers==4.51.1
tokenizers==0.21.1

# API and Environment
openai==1.72.0
python-dotenv==1.1.0

# Date processing
python-dateutil==2.9.0.post0

# Visualization
seaborn==0.13.2

# Development tools
ipykernel==6.29.5
pytest-cov==6.2.1

# Optional packages (for specific extractors)
# Uncomment if using specific features:
# medcat==1.14.2  # For RelCAT extractor
# spacy==3.7.5  # For spaCy-based NLP processing
# date-spacy==0.0.1  # For date extraction in notebooks
# faker==37.1.0  # For synthetic data generation
